perf: improve boot time by only traversing once
* refactor: only traverse once

The traverse module already recursively walks every node, this includes
nested objects and arrays. This makes the second traverse unnecessary and
increases the execution time significantly.

* refactor: skip enums and messages

By skipping objects with the key type, this should skip enums and
messages. Recursively iterating on enums and messages are unnecessary
and expensive as they contain the file descriptor protos which is a
buffer.

* refactor: remove usage of _.keys

* refactor: use single assignment to create object

* refactor: remove pick and reduce assignments

* docs: describe behaviour of call